+++ src/translations/de/buttons.php
@@ -3,4 +3,7 @@
 return [
     'Customizable Radio Buttons' => 'Individuelle Radio-Buttons',
     'Button group' => 'Schalterkonfiguration',
+    'Default value' => 'Standardwert',
+    'The button group from the config file to use.' => 'Die Button-Gruppe aus der Konfigurationsdatei die verwendet werden soll.',
+    'The default value of the field. If empty, no button will be preselected.' => 'Der Standardwert des Feldes. Ist dieser leer, wird kein Button vorausgewählt.'
 ];