+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Customizable Radio Buttons Field Changelog
 
+## 1.0.3 - 2021-09-16
+
+### Added
+
+- Add a simple way to set a default value to every field.
+- Add some instructions to the field settings.
+
 ## 1.0.2 - 2021-09-01
 
 ### Fixed
+
 - Add wrap and gap to the buttons (thanks to @emsuiko).
 - Fixed background color of icon.
 - Fixed a bug when two or more same configurations were used in one tab/ matrix block. 
@@ -10,6 +18,7 @@
 ## 1.0.1 - 2019-09-15
 
 ### Fixed
+
 - Publishing images to cpresources.
 
 ## 1.0.0 - 2019-04-06